requirements.txt
fastapi
uvicorn[standard]
python-multipart
paddleocr
paddlepaddle
pytesseract
Pillow
pillow-heif
pdf2image
deskew
opencv-python-headless
transformers
datasets
yfinance
requests
filelock
slowapi
faker
reportlab
pydantic
python-dotenv
feedparser
apscheduler
pytest
httpx
